new_plan/requirements.txt
# =============================================================================
# new_plan 依赖清单
# =============================================================================
# 最小集（可跑所有 --smoke）:
torch>=2.0
numpy>=1.23
pyyaml>=6.0

# LSTM1 训练/数据生成/可视化:
pandas>=1.5
matplotlib>=3.7

# ONNX 导出:
onnx>=1.14
# onnxruntime>=1.15   # 若需要 Python 端对 ONNX 做验证
